# src/main/resources/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Product API",
    "description": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
...
    "/api/v1/product/prices/find/one": {
      "get": {
        "tags": [
          "price-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET",
        "parameters": [
          {
            "name": "priceId",
            "in": "query",
            "description": "priceId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/products/add/one": {
      "post": {
        "tags": [
          "product-controller"
        ],
        "summary": "createProduct",
        "operationId": "createProductUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/products/delete/one": {
      "delete": {
        "tags": [
          "product-controller"
        ],
        "summary": "deleteById",
        "operationId": "deleteByIdUsingDELETE",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "description": "productId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/products/delete/one/permanent": {
      "delete": {
        "tags": [
          "product-controller"
        ],
        "summary": "deleteByIdPermanently",
        "operationId": "deleteByIdPermanentlyUsingDELETE",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "description": "productId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/products/edit/one": {
      "put": {
        "tags": [
          "product-controller"
        ],
        "summary": "updateProduct",
        "operationId": "updateProductUsingPUT",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "description": "productId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/products/exists/one": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "existsById",
        "operationId": "existsByIdUsingGET",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "description": "productId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/products/exists/unique": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "existsByUniqueFields",
        "operationId": "existsByUniqueFieldsUsingGET",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "description": "storeId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/products/get/all": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "findAllProducts",
        "operationId": "findAllProductsUsingGET",
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "description": "pageNo",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/products/get/one": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "findProductById",
        "operationId": "findProductByIdUsingGET",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "description": "productId",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateCategoryRequest": {
        "title": "CreateCategoryRequest",
        "type": "object",
        "properties": {
          "icon": {
            "type": "string",
            "format": "byte"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreateIngredientRequest": {
        "title": "CreateIngredientRequest",
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "format": "bigdecimal"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreatePriceRequest": {
        "title": "CreatePriceRequest",
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "bigdecimal"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreateProductRequest": {
        "title": "CreateProductRequest",
        "type": "object",
        "properties": {
          "categories": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCategoryRequest"
            }
          },
          "defaultImageId": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "etrackingId": {
            "type": "string"
          },
          "ingredients": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateIngredientRequest"
            }
          },
          "isFeatured": {
            "type": "boolean"
          },
          "limitPerOrder": {
            "type": "integer",
            "format": "int32"
          },
          "prices": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePriceRequest"
            }
          },
          "storeId": {
            "type": "integer",
            "format": "int64"
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTagRequest"
            }
          },
          "taxes": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTaxRequest"
            }
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "COMPLEX",
              "GROUPED",
              "SIMPLE"
            ]
          }
        }
      },
      "CreateTagRequest": {
        "title": "CreateTagRequest",
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
      "CreateTaxRequest": {
        "title": "CreateTaxRequest",
        "type": "object",
        "properties": {
          "isDefault": {
            "type": "boolean"
          },
          "percentage": {
            "type": "number",
            "format": "double"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ResponseEntity": {
        "title": "ResponseEntity",
        "type": "object",
        "properties": {
          "body": {
            "type": "object"
          },
          "statusCode": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "ALREADY_REPORTED",
              "BAD_GATEWAY",
              "BAD_REQUEST",
              "BANDWIDTH_LIMIT_EXCEEDED",
              "CHECKPOINT",
              "CONFLICT",
              "CONTINUE",
              "CREATED",
              "DESTINATION_LOCKED",
              "EXPECTATION_FAILED",
              "FAILED_DEPENDENCY",
              "FORBIDDEN",
              "FOUND",
              "GATEWAY_TIMEOUT",
              "GONE",
              "HTTP_VERSION_NOT_SUPPORTED",
              "IM_USED",
              "INSUFFICIENT_SPACE_ON_RESOURCE",
              "INSUFFICIENT_STORAGE",
              "INTERNAL_SERVER_ERROR",
              "I_AM_A_TEAPOT",
              "LENGTH_REQUIRED",
              "LOCKED",
              "LOOP_DETECTED",
              "METHOD_FAILURE",
              "METHOD_NOT_ALLOWED",
              "MOVED_PERMANENTLY",
              "MOVED_TEMPORARILY",
              "MULTIPLE_CHOICES",
              "MULTI_STATUS",
              "NETWORK_AUTHENTICATION_REQUIRED",
              "NON_AUTHORITATIVE_INFORMATION",
              "NOT_ACCEPTABLE",
              "NOT_EXTENDED",
              "NOT_FOUND",
              "NOT_IMPLEMENTED",
              "NOT_MODIFIED",
              "NO_CONTENT",
              "OK",
              "PARTIAL_CONTENT",
              "PAYLOAD_TOO_LARGE",
              "PAYMENT_REQUIRED",
              "PERMANENT_REDIRECT",
              "PRECONDITION_FAILED",
              "PRECONDITION_REQUIRED",
              "PROCESSING",
              "PROXY_AUTHENTICATION_REQUIRED",
              "REQUESTED_RANGE_NOT_SATISFIABLE",
              "REQUEST_ENTITY_TOO_LARGE",
              "REQUEST_HEADER_FIELDS_TOO_LARGE",
              "REQUEST_TIMEOUT",
              "REQUEST_URI_TOO_LONG",
              "RESET_CONTENT",
              "SEE_OTHER",
              "SERVICE_UNAVAILABLE",
              "SWITCHING_PROTOCOLS",
              "TEMPORARY_REDIRECT",
              "TOO_EARLY",
              "TOO_MANY_REQUESTS",
              "UNAUTHORIZED",
              "UNAVAILABLE_FOR_LEGAL_REASONS",
              "UNPROCESSABLE_ENTITY",
              "UNSUPPORTED_MEDIA_TYPE",
              "UPGRADE_REQUIRED",
              "URI_TOO_LONG",
              "USE_PROXY",
              "VARIANT_ALSO_NEGOTIATES"
            ]
          },
          "statusCodeValue": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UpdateCategoryRequest": {
        "title": "UpdateCategoryRequest",
        "type": "object",
        "properties": {
          "icon": {
            "type": "string",
            "format": "byte"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "UpdateIngredientRequest": {
        "title": "UpdateIngredientRequest",
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "format": "bigdecimal"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "UpdatePriceRequest": {
        "title": "UpdatePriceRequest",
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "bigdecimal"
          },
          "currency": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "isDefault": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "UpdateProductRequest": {
        "title": "UpdateProductRequest",
        "type": "object",
        "properties": {
          "categories": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateCategoryRequest"
            }
          },
          "defaultImageId": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "etrackingId": {
            "type": "string"
          },
          "ingredients": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateIngredientRequest"
            }
          },
          "isFeatured": {
            "type": "boolean"
          },
          "limitPerOrder": {
            "type": "integer",
            "format": "int32"
          },
          "prices": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdatePriceRequest"
            }
          },
          "storeId": {
            "type": "integer",
            "format": "int64"
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateTagRequest"
            }
          },
          "taxes": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateTaxRequest"
            }
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "COMPLEX",
              "GROUPED",
              "SIMPLE"
            ]
          }
        }
      },
      "UpdateTagRequest": {
        "title": "UpdateTagRequest",
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
      "UpdateTaxRequest": {
        "title": "UpdateTaxRequest",
        "type": "object",
        "properties": {
          "isDefault": {
            "type": "boolean"
          },
          "percentage": {
            "type": "number",
            "format": "double"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    }
  }
}